ncnews-17.A: refactor models and controllers into own files
