+++ api/controllers/topic-contoller.js
@@ -1,0 +1,11 @@
+const { fetchTopics } = require("../models/topic-model");
+
+exports.getTopics = (req, res, next) => {
+    fetchTopics()
+      .then((topics) => {
+        res.status(200).send(topics);
+      })
+      .catch((err) => {
+        next(err);
+      });
+};
