+++ api/models/endpoint-model.js
@@ -1,0 +1,10 @@
+const endpoints = require("../../endpoints.json");
+
+exports.fetchEndpoints = async () => {
+  try {
+    const returnedEndPoints = await endpoints;
+    return returnedEndPoints;
+  } catch (err) {
+    throw err;
+  }
+};
