+++ api/routes/article-router.js
@@ -1,19 +1,23 @@
-const { getArticles, getArticleById, patchArticleVotesByArticleId, getCommentsByArticleId, postCommentByArticleId } = require('../controllers/controllers');
+const {
+  getArticles,
+  getArticleById,
+  patchArticleVotesByArticleId,
+  getCommentsByArticleId,
+  postCommentByArticleId,
+} = require("../controllers/article-controller");
 
-const articleRouter = require('express').Router();
+const articleRouter = require("express").Router();
+
+articleRouter.route("/").get(getArticles);
 
 articleRouter
-    .route('/')
-    .get(getArticles)
+  .route("/:article_id")
+  .get(getArticleById)
+  .patch(patchArticleVotesByArticleId);
 
 articleRouter
-    .route('/:article_id')
-    .get(getArticleById)
-    .patch(patchArticleVotesByArticleId)
-
-articleRouter
-    .route('/:article_id/comments')
-    .get(getCommentsByArticleId)
-    .post(postCommentByArticleId)
+  .route("/:article_id/comments")
+  .get(getCommentsByArticleId)
+  .post(postCommentByArticleId);
 
 module.exports = articleRouter;
