+++ api/routes/user-router.js
@@ -1,9 +1,13 @@
-const { getUsers } = require('../controllers/controllers');
+const { getUsers } = require('../controllers/user-controller');
 
 const userRouter = require('express').Router();
 
 userRouter
     .route('/')
     .get(getUsers)
+
+userRouter
+    .route('/:username')
+    .get()
 
 module.exports = userRouter;
