+++ api/utils/input-validators.js
@@ -1,9 +1,8 @@
 exports.isValidArticleColumn = (sort_by) => {
-    const columns = ["title", "topic", "author", "body", "created_at", "votes"]
-    return columns.includes(sort_by);
-
-}
-exports.isValidOrder = (order) =>{
-    const columns = ["asc", "desc"]
-    return columns.includes(order);
-}
+  const columns = ["title", "topic", "author", "body", "created_at", "votes"];
+  return columns.includes(sort_by);
+};
+exports.isValidOrder = (order) => {
+  const columns = ["asc", "desc"];
+  return columns.includes(order);
+};
